+++ DawnLib.Dusk/src/API/DuskModContent.cs
@@ -7,4 +7,5 @@ public static class DuskModContent
     public static Registry<DuskAchievementDefinition> Achievements = new();
     public static Registry<DuskVehicleDefinition> Vehicles = new();
     public static Registry<DuskEntityReplacementDefinition> EntityReplacements = new();
+    public static Registry<DuskNamespacedObjectDefinition> NamespacedObjects = new();
 }