+++ DawnLib.Dusk/src/DuskPlugin.cs
@@ -37,6 +37,7 @@ public class DuskPlugin : BaseUnityPlugin
         CommitKeyToSave.Init();
         if (!DawnConfig.VanillaCompatibility.Value)
         {
+            NamespacedSpawnSyncedObject.Init();
             NetworkerPatch.Init();
             VehicleRegistrationPatch.Init();
         }